Adds the RecordIdentifier and move methods out of the NamesResolver
I felt like the NamesResolver class could be improved, so I started
digging the Rails code base and found out that they have a Name class
which I really liked. So I brought that same approach here.
Unfortunatelly, we cannot add that to the model because we have to
allow broadcasting models that do not use the Broadcasts trait. But that
name class is still useful, I think.

I also added a "RecordIdentifier" class to handle the domId and domClass
behavior. It felt cleaner. I wasn't able to find a good place for the
other methods currently in the NamesResolver, so I kept those there.
Maybe they belong there?