Adds a TurboStreamBroadcasts trait to streamline events
I think there is a way to have a single broadcast event instead of 3,
but I wasn't able to find an abstraction yet. I'm thinking of keeping
the 3 events and add a listener that would call the broadcast() helper
function using a single broadcasting event. I'll give it another attempt soon.