Allow passing a string to the shorthand methods and fix before/after
The before/after shorthand were working weirdly, we had to pass the
model and the target. It didn't make much sense. Now, we just need
to pass the model to it. This means the shorthand only indicates the
target and action. You'll need an extra call to pass the view/partial
but the overall usage of the helper makes a lot more sense to me now.

All shorthand methods allow passing an inline content which can be
either a string or an instance of HtmlString, since this will render
with the double curly-braces (escaping) in the underlying Blade
template.